update for recent versions of gdbm
Recent versions of gdbm don't have a global gdbm_errno, it's a
thread-local variable. the address of that variable can be retrieved
with gdbm_errno_location(). gdbm_errno in C is #defined as
*gdbm_errno_location.

So remove the gdbm_errno export and add gdbm_errno_location.

While we're at it, gdbm_strerror() is now a const function, so update
that as well.

Update version to 0.3.